refactor: split GitTabHelper into GitTabDataBuilder + GitTabView, migrate Views namespace (#779)
- Split static GitTabHelper into ViewBase subclass (GitTabView) and data helper (GitTabDataBuilder)
- Move Views/ to Apps/Views/ and unify namespace to Ivy.Tendril.Apps.Views
- Fix inconsistent hash truncation (9 → 7 chars) in ResolveParentInfo
- Services now resolved via UseService<T>() instead of parameter passing
