+++ src/Ivy.Tendril/Apps/DashboardApp.cs
@@ -1,6 +1,6 @@
+using Ivy.Tendril.Apps.Views;
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
-using Ivy.Tendril.Views;
 
 namespace Ivy.Tendril.Apps;
