+++ src/Ivy.Tendril/Apps/Icebox/ContentView.cs
@@ -1,8 +1,8 @@
 using Ivy.Core;
 using Ivy.Tendril.Apps.Icebox.Dialogs;
 using Ivy.Tendril.Apps.Plans;
-using Ivy.Tendril.Views;
-using Ivy.Tendril.Views.Sheets;
+using Ivy.Tendril.Apps.Views;
+using Ivy.Tendril.Apps.Views.Sheets;
 using Ivy.Tendril.Models;
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
