+++ src/Ivy.Tendril/Apps/JobsApp.cs
@@ -1,8 +1,8 @@
 using System.Reactive.Linq;
 using Ivy.Tendril.Apps.Jobs;
+using Ivy.Tendril.Apps.Views.Sheets;
 using Ivy.Tendril.Helpers;
 using Ivy.Tendril.Services;
-using Ivy.Tendril.Views.Sheets;
 
 namespace Ivy.Tendril.Apps;
