+++ src/Ivy.Tendril/Apps/Plans/PlanTabView.cs
@@ -1,8 +1,8 @@
 using Ivy.Tendril.Models;
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
-using Ivy.Tendril.Views.Sheets;
 using Ivy.Core;
+using Ivy.Tendril.Apps.Views.Sheets;
 
 namespace Ivy.Tendril.Apps.Plans;
