+++ src/Ivy.Tendril/Apps/Recommendations/SidebarView.cs
@@ -1,6 +1,6 @@
 using Ivy.Tendril.Services;
 using Ivy.Tendril.Helpers;
-using Ivy.Tendril.Views;
+using Ivy.Tendril.Apps.Views;
 
 namespace Ivy.Tendril.Apps.Recommendations;
