+++ src/Ivy.Tendril/Apps/Views/NoContentView.cs
@@ -1,10 +1,10 @@
-namespace Ivy.Tendril.Views;
+namespace Ivy.Tendril.Apps.Views;
 
 public class NoContentView(string title, string description, object? cta = null) : ViewBase
 {
     public override object Build()
     {
-        var layout = Layout.Vertical().AlignContent(Align.Center).Height(Size.Full()).Gap(4).Padding(8)
+        var layout = Layout.Vertical().AlignContent(Align.Center).Height(Size.Full()).Padding(8)
                      | Text.H3(title)
                      | Text.Muted(description);
