+++ src/Ivy.Tendril/Apps/Views/Sheets/CommitDetailSheet.cs
@@ -1,9 +1,8 @@
-using Ivy.Core;
-using Ivy.Tendril.Models;
 using Ivy.Tendril.Helpers;
-
+using Ivy.Tendril.Models;
 using Ivy.Tendril.Services;
-namespace Ivy.Tendril.Views.Sheets;
+
+namespace Ivy.Tendril.Apps.Views.Sheets;
 
 public class CommitDetailSheet(
     IState<string?> openCommit,
