+++ src/Ivy.Tendril/Apps/Views/Sheets/FileSheet.cs
@@ -2,7 +2,7 @@ using System.Diagnostics;
 using Ivy.Tendril.Helpers;
 using Ivy.Tendril.Services;
 
-namespace Ivy.Tendril.Views.Sheets;
+namespace Ivy.Tendril.Apps.Views.Sheets;
 
 public class FileSheet(
     IState<string?> openFile,
